Bug with non existing alias file for compiler modules fixed.
refs #58, fixes #58.

+++ actions/CLangLinker.rb
@@ -20,23 +20,27 @@ class CLangLinker
 		
 	end
 	
-	def ParsePathAliases(path)
-		
-		File.open(path) do |f|
-		
-			f.each_line do |l|
-			
-				name, value = l.split('=')
-				
-				@pathAliases[name] = value.chop!
-				
 
-			
-			end
+	def ParsePathAliases(path)
+
+		if File.exist? path
 		
+			File.open(path) do |f|
+		
+				f.each_line do |l|
+			
+					name, value = l.split('=')
+				
+					@pathAliases[name] = value.chop!
+				
+				end
+		
+			end
+
 		end
 	
 	end
+	
 	
 	def Keys(rawStr)
