+++ actions/MSCompiler.rb
@@ -13,23 +13,26 @@ class MSCompiler
 	end
 	
 	def ParsePathAliases(path)
-		
-		File.open(path) do |f|
-		
-			f.each_line do |l|
-			
-				name, value = l.split('=')
-				
-				@pathAliases[name] = value.chop!
-				
 
-			
-			end
+		if File.exist? path
 		
+			File.open(path) do |f|
+		
+				f.each_line do |l|
+			
+					name, value = l.split('=')
+				
+					@pathAliases[name] = value.chop!
+				
+				end
+		
+			end
+
 		end
 	
 	end
 	
+
 	def TouchDir(name)
 	
 		begin
